requirements.txt
gradio==5.44.1


torch>=2.1.0
transformers>=4.35.0
accelerate>=0.26.0

huggingface-hub

tokenizers

numpy


tts


python-dotenv
gtts
pydub